Fixed Time Machine backup function
Amended Time Machine backup function to fix an error that reported a TM db had been found when TM may not have been configured at all/there may not have been a backup partition. 
Revision to function is as below:
First TM check is now if the Backup partition exists, if not the keychain is backed up. If one does exist the latest backup date is compared to todays date, if they match the keychain is deleted and if they don't the keychain is backed up